Handle small differences between language servers
Try to be more resilient against small differences between language
servers.  Avoid errors when:
- There is a (non-error) response, but no `result` key
- There is a (non-error) response, but the `result` key is empty

In addition most(?) language servers return a result of the form:
result = { { targetRange = {...}, ... } }
However some (at least FsAutocomplete) returns a result of the form:
result = { range = {...}, ... }  -- note the lack of enclosing table

The difference in `range` vs `targetRange` is handled in
`tags.formatting`, but this does not expect the missing enclosing table.
Add it when we detect this situation.